// src/app/models/Comic.ts
export interface Comic {
    id?: number;
    nameComic?: string;
    editorial?: string;
    topic?: string;
    collections?: string;
    pages?: number;
    price?: number;
    autors?: string;
    color?: boolean;
    lastSaleDate?: Date;
    status?: boolean;
    amount?: number;
    url_img?: string;
    exitoso?: string;
    mensajeEjecucion?: string;
    createdAt?: Date;
    updatedAt?: Date;
    _id?: number;
    __v?: string;
}